Move table of contents types into lib
