+++ packages/lesswrong/components/common/SidebarsWrapper.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useMemo, useState } from 'react';
-import { Components, registerComponent } from '../../lib/vulcan-lib';
-import type { ToCData, ToCSection } from '../../server/tableOfContents';
+import { registerComponent } from '../../lib/vulcan-lib';
+import type { ToCData } from '../../lib/tableOfContents';
 
 // Context used to share a reference used to share the table of contents
 // between the ToC itself, and the Header. The Header uses the ToC to change
