+++ packages/lesswrong/components/common/TabNavigationMenu/NavigationDrawer.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { registerComponent, Components } from '../../../lib/vulcan-lib';
 import SwipeableDrawer from '@material-ui/core/SwipeableDrawer';
 import classNames from 'classnames';
-import type { ToCData } from '../../../server/tableOfContents';
+import type { ToCData } from '../../../lib/tableOfContents';
 
 const styles = (theme: ThemeType): JssStyles => ({
   paperWithoutToC: {
