+++ packages/lesswrong/components/posts/TableOfContents/TableOfContentsList.tsx
@@ -3,7 +3,7 @@ import { Components, registerComponent } from '../../../lib/vulcan-lib';
 import withErrorBoundary from '../../common/withErrorBoundary'
 import { isServer } from '../../../lib/executionEnvironment';
 import { useLocation, useNavigation } from '../../../lib/routeUtil';
-import type { ToCData, ToCSection } from '../../../server/tableOfContents';
+import type { ToCData, ToCSection } from '../../../lib/tableOfContents';
 import qs from 'qs'
 import isEmpty from 'lodash/isEmpty';
 import filter from 'lodash/filter';
