+++ packages/lesswrong/components/sequences/CollectionTableOfContents.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { getBookAnchor } from '../../lib/collections/books/helpers';
 import { registerComponent, Components } from '../../lib/vulcan-lib';
-import type { ToCSection } from '../../server/tableOfContents';
+import type { ToCSection } from '../../lib/tableOfContents';
 import { commentBodyStyles } from '../../themes/stylePiping';
 
 const styles = (theme: ThemeType): JssStyles => ({
